# backend/api.yaml
openapi: 3.0.2
servers:
  - url: http://127.0.0.1:3000
info:
  description: This is API Doc for Angular+ NodeJS Bulletinboard
  title: Swagger API Doc for Gigger
  version: 1.0.0
...
paths:
...
  /api/logout:
    get:
      tags:
        - Authentication
      summary: Logs out current logged in user session
      security:
        - {}
      description: ""
      operationId: logoutUser
      parameters:
        - name: userId
          in: header
          description: ID of login UserId
          required: true
          schema:
            type: string
      responses:
        default:
          description: Logout Successfully.
...
  /api/v1/posts:
    get:
      tags:
        - Posts
      summary: Get all Posts
      description: This can only be done by the logged in user.
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: PageNumber 
          schema:
            type: number
        - name: size
          in: query
          description: Number of users that want to show per Page
          schema:
            type: number
        - name: userId
          in: header
          description: ID of login UserId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Returns all Posts
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    post:
      parameters:
        - name: userId
          in: header
          description: ID of login UserId
          required: true
          schema:
            type: string
      tags:
        - Posts
      summary: Create Post
      description: This can only be done by the logged in user.
      security:
        - bearerAuth: []
      operationId: createPost
      requestBody:
        description: Created post
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                caption:
                  type: string
                  example: "Pink Floyd Live"
                music:
                  type: json
                  example: [1, 2]
                address:
                  type: string
                  example: "No(35) Green Street"
                advertisementFormat:
                  type: string
                  example: "1"
                searchResult:
                  type: string
                  example: "Artists"
                giglistClassifieds:
                  type: string
                  example: "Musicians"
                targetAudience:
                  type: string
                  example: "Auto"
                privateContent:
                  type: boolean
                  example: true
                memberShipContent:
                  type: boolean
                  example: true
                forMyFollowersOnly:
                  type: boolean
                  example: true
                video:
                  type: string
                  format: binary
          application/octet-stream:
            schema:
              type: object
              properties:
                caption:
                  type: string
                  example: "Pink Floyd Live"
                music:
                  type: json
                  example: [1, 2]
                address:
                  type: string
                  example: "No(35) Green Street"
                advertisementFormat:
                  type: string
                  example: "1"
                searchResult:
                  type: string
                  example: "Artists"
                giglistClassifieds:
                  type: string
                  example: "Musicians"
                targetAudience:
                  type: string
                  example: "Auto"
                privateContent:
                  type: boolean
                  example: true
                memberShipContent:
                  type: boolean
                  example: true
                forMyFollowersOnly:
                  type: boolean
                  example: true
                video:
                  type: string
                  example: "1.mp4"
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                caption:
                  type: string
                  example: "Pink Floyd Live"
                music:
                  type: json
                  example: [1, 2]
                address:
                  type: string
                  example: "No(35) Green Street"
                advertisementFormat:
                  type: string
                  example: "1"
                searchResult:
                  type: string
                  example: "Artists"
                giglistClassifieds:
                  type: string
                  example: "Musicians"
                targetAudience:
                  type: string
                  example: "Auto"
                privateContent:
                  type: boolean
                  example: true
                memberShipContent:
                  type: boolean
                  example: true
                forMyFollowersOnly:
                  type: boolean
                  example: true
                video:
                  type: string
                  format: binary
      responses:
        "200":
          description: Created event Successfully.
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /api/v1/posts/{id}:
    get:
      tags:
        - Posts
      summary: Get post by ID
      description: This can only be done by the logged in user.
      security:
        - bearerAuth: []
      operationId: getPostById
      parameters:
        - name: id
          in: path
          description: ID of post to return
          required: true
          schema:
            type: string
        - name: userId
          in: header
          description: ID of login UserId
          required: true
          schema:
            type: string
      responses:
          "200":
            description: Returns all Posts
          "401":
            $ref: "#/components/responses/UnauthorizedError"
          "404":
            description: "Post is not found"
  /api/v1/posts/update/{id}:
    post:
      tags:
        - Posts
      summary: Update a post
      description: This can only be done by the logged in user.
      security:
        - bearerAuth: []
      operationId: updatePostById
      parameters:
        - name: id
          in: path
          description: ID of post to return
          required: true
          schema:
            type: string
        - name: userId
          in: header
          description: ID of login UserId
          required: true
          schema:
            type: string
      operationId: updatePost
      requestBody:
        description: Updated post
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                caption:
                  type: string
                  example: "Pink Floyd Live"
                music:
                  type: json
                  example: [1, 2]
                address:
                  type: string
                  example: "No(35) Green Street"
                advertisementFormat:
                  type: string
                  example: "1"
                searchResult:
                  type: string
                  example: "Artists"
                giglistClassifieds:
                  type: string
                  example: "Musicians"
                targetAudience:
                  type: string
                  example: "Auto"
                privateContent:
                  type: boolean
                  example: true
                memberShipContent:
                  type: boolean
                  example: true
                forMyFollowersOnly:
                  type: boolean
                  example: true
                video:
                  type: string
                  format: binary
          application/octet-stream:
            schema:
              type: object
              properties:
                caption:
                  type: string
                  example: "Pink Floyd Live"
                music:
                  type: json
                  example: [1, 2]
                address:
                  type: string
                  example: "No(35) Green Street"
                advertisementFormat:
                  type: string
                  example: "1"
                searchResult:
                  type: string
                  example: "Artists"
                giglistClassifieds:
                  type: string
                  example: "Musicians"
                targetAudience:
                  type: string
                  example: "Auto"
                privateContent:
                  type: boolean
                  example: true
                memberShipContent:
                  type: boolean
                  example: true
                forMyFollowersOnly:
                  type: boolean
                  example: true
                video:
                  type: string
                  format: binary
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                caption:
                  type: string
                  example: "Pink Floyd Live"
                music:
                  type: json
                  example: [1, 2]
                address:
                  type: string
                  example: "No(35) Green Street"
                advertisementFormat:
                  type: string
                  example: "1"
                searchResult:
                  type: string
                  example: "Artists"
                giglistClassifieds:
                  type: string
                  example: "Musicians"
                targetAudience:
                  type: string
                  example: "Auto"
                privateContent:
                  type: boolean
                  example: true
                memberShipContent:
                  type: boolean
                  example: true
                forMyFollowersOnly:
                  type: boolean
                  example: true
                video:
                  type: string
                  format: binary
      responses:
        "200":
          description: Post is updated successfully.
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          description: Post is not found.
...
components:
...
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT # optional, arbitrary value for documentation purposes
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid

# 2) Apply the security globally to all operations
security:
  - bearerAuth: []